refactor(tools)!: move /v1/tools route to /v1/admin/tools
Tools are admin-controlled resources configured at the distribution level,
not user-scoped. Move the tools listing endpoint under /v1/admin for
consistency with the connectors move in #5659. The Admin protocol now
includes list_tools, and AdminImpl delegates to the ToolGroups routing
table implementation.

BREAKING CHANGE: GET /v1/tools is now served at GET /v1/admin/tools.
