+++ src/ogx_api/admin/api.py
@@ -6,6 +6,8 @@
 
 from typing import Protocol, runtime_checkable
 
+from ogx_api.tools.models import ListToolDefsResponse, ListToolsRequest
+
 from .models import (
     HealthInfo,
     InspectProviderRequest,
@@ -68,5 +70,15 @@ class Admin(Protocol):
         Get the version of the service.
 
         :returns: Version information containing the service version number.
+        """
+        ...
+
+    async def list_tools(self, request: ListToolsRequest) -> ListToolDefsResponse:
+        """List tools.
+
+        List tools with optional tool group filter.
+
+        :param request: Request containing optional filter parameters
+        :returns: A ListToolDefsResponse containing available tool definitions.
         """
         ...
